Check/Update SummarizedBenchmark

Function to update or check status of SummarizedBenchmark results.

If only a SummarizedBenchmark object is specified, the function will check whether `func`, `param`, `meta`, `post` or the `pkg_vers` of the methods in the BenchDesign stored with the SummarizedBenchmark do not match values stored in the colData. By default, no methods will be executed to update results. To actually execute updates, set dryrun = FALSE.

If a BenchDesign object is specified in addition to a SummarizedBenchmark object, the function will check which methods in the new BenchDesign need to be executed to update the SummarizedBenchmark results. Again, by default, no methods will be executed unless dryrun = FLASE is specified.

Unless reuseParams = FALSE is specified, the parameters of the last execution session stored in the the colData of the SummarizedBenchmark object will be used.

updateBench(sb, bd = NULL, dryrun = TRUE, version = FALSE,
  keepAll = TRUE, reuseParams = TRUE, ...)

Arguments

sb
a SummarizedBenchmark object

bd
a BenchDesign object

dryrun
logical whether to just print description of what would be updated rather than actually running any methods. (default = TRUE)

version
logical whether to re-run methods with only package version differences. (default = FALSE)

keepAll
logical whether to keep methods run in original SummarizedBenchmark but not in new BenchDesign. Only used if bd is not NULL. (default = TRUE)

reuseParams
logical whether to reuse parameters from buildBench call used to create SummarizedBenchmark object (if available). Directly specified buildBench parameters still take precedence. (default = TRUE)

...
optional parameters to pass to buildBench.

Value

SumamrizedBenchmark object.

See also

Examples

## load example SummarizedBenchmark object
data(allSB)

#> Warning: data set ‘allSB’ not found

sb <- allSB[[1]]

## check if results are out of date
updateBench(sb)

#> Update SummarizedBenchmark (dryrun) --------------------------- 
#>   benchmark data: unchanged (full data missing)
#>     MD5 hash: 6a54d3a70bb48f54435c56fa8401adc4 
#>     names: ??
#>   benchmark methods:
#>     |              | Need to  |           Outdated            |
#>     | Method       | (Re)Run  |  Func Param  Meta  Post  Vers |
#>     | kallisto-... |       Y  |     N     N     Y     N     N |
#>     | kallisto-... |       Y  |     N     N     Y     N     N |
#>     | sailfish-... |       Y  |     N     N     Y     N     N |
#>     | salmon-de... |       Y  |     N     N     Y     N     N |
#>     | salmon-gc... |       Y  |     N     N     Y     N     N |

## modify BenchDesign
bd <- BenchDesign(sb)
bd <- dropMethod(bd, "kallisto-default")

## check if results need to be updated with new BenchDesign
updateBench(sb, bd)

#> Update SummarizedBenchmark (dryrun) --------------------------- 
#>   benchmark data: unchanged (full data missing)
#>     MD5 hash: 6a54d3a70bb48f54435c56fa8401adc4 
#>     names: ??
#>   benchmark methods:
#>     |              | Need to  |           Outdated            |
#>     | Method       | (Re)Run  |  Func Param  Meta  Post  Vers |
#>     | kallisto-... |       Y  |     N     N     Y     N     N |
#>     | sailfish-... |       Y  |     N     N     Y     N     N |
#>     | salmon-de... |       Y  |     N     N     Y     N     N |
#>     | salmon-gc... |       Y  |     N     N     Y     N     N |
#>     | kallisto-... |       N  |     -     -     -     -     - |