SummarizedBenchmark class

Extension of the RangedSummarizedExperiment to store the output of different methods intended for the same purpose in a given dataset. For example, a differential expression analysis could be done using limma-voom, edgeR and DESeq2. The SummarizedBenchmark class provides a framework that is useful to store, benckmark and compare results.

Slots

performanceMetrics

A SimpleList of the same length as the number of assays containing performance functions to be compared with the ground truths.

BenchDesign

A BenchDesign originally used to generate the results in the object.

Examples

## loading the example data from iCOBRA
library(iCOBRA)

#> 
#> Attaching package: ‘iCOBRA’

#> The following objects are masked from ‘package:GenomicRanges’:
#> 
#>     score, score<-

#> The following objects are masked from ‘package:IRanges’:
#> 
#>     score, score<-

#> The following objects are masked from ‘package:BiocGenerics’:
#> 
#>     score, score<-

data(cobradata_example)

## a bit of data wrangling and reformatting
assays <- list(
    qvalue=cobradata_example@padj,
    logFC=cobradata_example@score )
assays[["qvalue"]]$DESeq2 <- p.adjust(cobradata_example@pval$DESeq2, method="BH")
groundTruth <- DataFrame( cobradata_example@truth[,c("status", "logFC")] )
colnames(groundTruth) <- names( assays )
colData <- DataFrame( method=colnames(assays[[1]]) )
groundTruth <- groundTruth[rownames(assays[[1]]),]

## constructing a SummarizedBenchmark object
sb <- SummarizedBenchmark(
    assays=assays, colData=colData,
    groundTruth=groundTruth )